data pulling is done/ app is basically done. Just need to add all of the pokemon to the database

+++ Pokedex/Model/PokemonData.swift
@@ -13,11 +13,11 @@ struct PokemonData: Decodable, Identifiable {
     let name: String
     let imageUrl: String
     let type: String
-    let weight: Int
-    let height: Int
+    let weight: Float
+    let height: Float
     let description: String
-    let attack: Int
-    let defense: Int
+    let attack: Float
+    let defense: Float
     let typeTwo: String
 }
